+++ dbt/models/staging/stg_sales.sql
@@ -8,7 +8,7 @@ renamed as (
         cast(customer_id as int64)  as client_id,
         cast(date_date as date)     as order_date,
         cast(products_id as int64)  as product_id,
-        cast(net_sales as float64)  as net_sales,
+        cast(net_sales as numeric)  as net_sales,
         cast(qty as int64)          as order_qty
     from source
 )
